SLE crawl function - working
I'm going to write the SLE crawl into a seperate function and remove it from the main code.

Liam also helped to write the function 'slideWindow' and 'pwelch_plot' tha thelp to analyze the frequency content of seizures, but it doesn't seem very accurate.

Kramay then helped me to understand what's going on from the theory side of signal processing and I realized it's too complex. I will simply count the number of spikes that occur in a seizure to infer the frequency.