Check point
Code is running smoothly and able to plot all epileptiform events.
Need to use this output to study the characteristics of seizures so we know how to design the classifier so it only selects the epileptiform events that I would consider to be an ictal event, and all others as IIEs.