checkpoint - V6.4
The Crawler is not working

I used only 1.5s of context to search for the base line.

I used power feature of the  absolute signal to search for the offset.

I was able to improve detection of offset. Notably, by better detecting the ictal events, I was able to better classify their features. This is better for the population statistics calculations.