7th Iieration - complete
plotEvent function is complete. I had to make a figure first in the script, then call the figure into the plotEvent function to plot all the events detected with their located spikes.

The script is running smoothly. Takes longer because it recreates the variables twice with the plotEvent function, but it's worth it because it keep the code neat and tidy.