checkpoint - crawler function
I've updated the crawler function so that now it will find the onset better (it may lie before the light pulse by a few ms). To see if a spike is light-triggered, I check to see if the peak of the spike comes after the light pulse).

Default use 3 s of context when searching for the offset of the spike