updated crawler.m
If the last spike is caused by a light-trigger, then I  take the precenting spike's location as the offset.

I don't scan the offset baseline for when the power decreases to threshold as the offset, because then it'll accidently search the spike triggered by the light pulse and use that as the offset.