updated frequencyanalysis
The scriptEpileptiformFrequencyWorkSpace.m has now been updated to plot the proper times and padding at the end and beginning so you can capture all the frequency associated with the epilpetiform event (those at the edges are now removed). I put green and red markers to identify in the LFP plot where the approximate onset and offset for the SLE are located (however because they were int64 adjusted, they may be slightly shifted). refer to the exports from detectionAlgorithm for exact onset and offset locations.