Updates to Crawler.m
Updates to Crawler now allow for detecting the delay between the stimulus onset and ictal event onset. Also, the interstimulus interval, which 'sandwiches' the ictal event, is calculated as well. These values can be used for circular statistics to see if the events are correlated to light according to Chang et al., 2018, Neurobiology of Disease.