Update crawler.m
Now it will detect delay to ictal event onset by subtracting the peak of the epileptiform event's first spike from light pulse onset.

It will not represent the delay to the ictal event onset, but it will be more informative for calculating the circular statistics and graphic the rose plots because sometimes the ictal event onset is detecft a few ms too early, and can make it seem like it's a spontaneous ictal event (when in fact, it was light-triggered). Unfortunately, this is a feature of how the algorithm tries to detecft the onset of epileptiform events using the power signiture and arbitary thresholding.