Update scriptFrequencyContent.m
This commit contains the working version of scriptFrequencyContent.m

It can sufficiently analyze files can report their frequency content as well as 'normalized' frequency content. It attempts to normalize the ictal event's frequency content by removing the frequencies that are found in a snippet of data with the lowest sigma in voltage activity. However, this is not a very effective method because the recording is contains many different frequency content, when you remove the dominiant frequency, another dominiant frequency appears (and so on and so forth).